% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{grid_density}
\alias{grid_density}
\title{Evaluate the multivariate normal density on a grid}
\usage{
grid_density(grid, theta_mean, theta_cov)
}
\arguments{
\item{grid}{Numeric matrix (N x D).}

\item{theta_mean}{Numeric vector of length D.}

\item{theta_cov}{Numeric D x D positive-definite matrix.}
}
\value{
Numeric vector of length N.
}
\description{
Evaluate the multivariate normal density on a grid
}
\keyword{internal}
