% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{make_grid}
\alias{make_grid}
\title{Build a quadrature grid over \verb{[-6, 6]^D}}
\usage{
make_grid(D, n_pts)
}
\arguments{
\item{D}{Integer number of dimensions.}

\item{n_pts}{Integer number of points per dimension.}
}
\value{
List with components \code{grid} (matrix, nrow = \code{n_pts^D}, ncol = D) and
\code{dx} (scalar step size).
}
\description{
Build a quadrature grid over \verb{[-6, 6]^D}
}
\keyword{internal}
